chore: remove playwright artifacts
