+++ frontend/package.json
@@ -9,19 +9,23 @@
     "preview": "vite preview",
     "format": "prettier -w src",
     "lint": "eslint src --ext .ts,.svelte",
-    "test": "playwright test"
+    "test": "vitest run",
+    "test:watch": "vitest"
   },
   "devDependencies": {
     "@sveltejs/kit": "^2.5.0",
     "@sveltejs/adapter-vercel": "^5.3.0",
-    "svelte": "^5.0.0",
-    "typescript": "^5.0.0",
-    "tailwindcss": "^3.3.0",
-    "postcss": "^8.4.0",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/svelte": "^4.0.4",
     "autoprefixer": "^10.4.0",
-    "prettier": "^3.2.5",
     "eslint": "^8.57.0",
-    "@playwright/test": "^1.42.1",
-    "vite": "^6.3.0"
+    "jsdom": "^24.0.0",
+    "postcss": "^8.4.0",
+    "prettier": "^3.2.5",
+    "svelte": "^5.0.0",
+    "tailwindcss": "^3.3.0",
+    "typescript": "^5.0.0",
+    "vite": "^6.3.0",
+    "vitest": "^1.5.0"
   }
 }
