--- frontend/tests/example.spec.ts
@@ -1,6 +1,0 @@
-import { test, expect } from '@playwright/test'
-
-test('home page has title', async ({ page }) => {
-  await page.goto('/')
-  await expect(page.locator('h1')).toHaveText('PO Drafter Chat Wizard')
-})
